jiujitsuInfo/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autopep8==2.3.1
beautifulsoup4==4.12.3
billiard==4.2.1
black==24.10.0
blis==1.0.1
boto3==1.36.21
botocore==1.36.21
cachetools==5.5.0
catalogue==2.0.10
celery==5.4.0
certifi==2024.8.30
cfn-flip==1.3.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.20.0
confection==0.1.5
cron-descriptor==1.4.5
cymem==2.0.8
Django==3.2.25
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-dialogflow==2.38.0
google-cloud-vision==3.8.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.14.0
hjson==3.1.0
httplib2==0.22.0
idna==3.10
ja_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/ja_core_news_sm-3.8.0/ja_core_news_sm-3.8.0-py3-none-any.whl#sha256=65693016b7cf33ffa845af3523a9c9b8d9aabae85dfc2e46731513a5e03def04
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kappa==0.6.0
kombu==5.4.2
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.1.1
nltk==3.9.1
numpy==2.0.2
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
placebo==0.9.0
platformdirs==4.3.6
playwright==1.49.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
selenium==4.27.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
SudachiDict-core==20241021
SudachiPy==0.6.8
text-unidecode==1.3
thinc==8.3.2
timedelta==2020.12.3
toml==0.10.2
tomli==2.1.0
tqdm==4.67.0
trio==0.27.0
trio-websocket==0.11.1
troposphere==4.9.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.32.1
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.2
zappa==0.59.0
whitenoise
